SonarQube issue cleanup across the reactor.
- Strip redundant `throws Exception` from heavy test files (RestLogger_Test,
  DefaultSwaps_Test, TomlWriter_Test, CsvParser_Test, SamlAssertionValidator_Builder_Test,
  ConfigClasspathStore_Test).
- Remove unnecessary casts (java:S1905) across the token-streaming parser/serializer
  role-interface family and related client code.
- Add by-design rule suppressions with one-line rationales (e.g. java:S1192 on the
  HTTP header beans).
- Fix a ConfigStore ThreadLocal cleanup (remove the LOADING ThreadLocal when the
  per-thread loading set empties).

+++ juneau-core/juneau-marshall/src/main/java/org/apache/juneau/marshall/json/JsonTokenWriter.java
@@ -446,7 +446,7 @@ public class JsonTokenWriter implements TokenWriter {
 	private void pushContext(int kind) {
 		if (depth < 32) {
 			// Two bits per level: 00 = root, 01 = object, 10 = array.
-			ctxBits |= ((long) kind & 0x3L) << (depth * 2);
+			ctxBits |= (kind & 0x3L) << (depth * 2);
 		} else {
 			if (ctxOverflow == null || depth - 32 >= ctxOverflow.length)
 				growOverflow();
